It seems to me like the regular expression on line 48 that's checking to see whether a stylesheet's HREF is external or not doesn't correctly identify "protocol-less" network-path references as being external. For instance, suppose I'm including some YUI CSS file from Google's CDN (and I want to fetch it using whatever protocol the page was delivered in), in my LINK tag, I would set the HREF to this: "//ajax.googleapis.com/ajax/libs/yui/2.9.0/build/base/base-min.css"
Respond.js erroneously thinks this is a local file and adds it to the requestQueue. When I include this on a page in IE7, I see an "Access is denied" error on the page.

I believe my updated regular expression correctly treats network-path references as external while still treating everything else the same (though I would love to have other people take a look at it and make sure).