Fix replaceStartIndex === -1 replace bug.
As a result of rebasing the `RESPOND_REPLACE_STYLES` option code onto
1c8ce14b10cf636ed72a9e0ccffedb71143e2e69 (after previously being on 1.4.2), the
replace index was being calculated as -1, meaning the query could not be found
in the total styles. This was because comments are now being stripped.

To fix that, the stripping of comments and keyframes from styles are done
immediately in the `translate` function, and then stored with their replaced
urls if `RESPOND_REPLACE_STYLES === true`.